requirements.txt
cryptography
numpy
sklearn
transformers
datasets
torchvision
pyfiglet
tqdm

